# Performance improvement in createAstro function