@LeanderG 2024-04-12
@lilnasy
I used `pnpm run benchmark`in the Astro repository.

I also tried out the benchmark you linked.
There is only an improvement when rendering many children. Try extracting
```
<tr>
  <td>{entry.id}</td>
  <td>{entry.name}</td>
</tr>
```

into a component.